GUI/requirements.txt
numpy==2.2.0
opencv-python==4.10.0.84
paho-mqtt==2.1.0
pyqtgraph==0.13.7
PySide6==6.8.1
PySide6_Addons==6.8.1
PySide6_Essentials==6.8.1
shiboken6==6.8.1
